requirements.txt
requests==2.22.0
pandas==0.25.3
google-cloud-storage==1.23.0
pandas-gbq==0.12.0
google-cloud-bigquery==1.22.0